requirements.txt
altair==5.5.0
altgraph==0.17.4
anyio==4.9.0
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
customtkinter==5.2.2
cycler==0.12.1
Cython==3.0.12
darkdetect==0.8.0
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
et_xmlfile==2.0.0
Flask==3.0.3
fonttools==4.56.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.70.0
gspread==6.1.4
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
MouseInfo==0.1.3
narwhals==1.27.1
nest-asyncio==1.6.0
numpy==2.2.3
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pillow==11.1.0
plotly==6.0.0
proto-plus==1.26.1
protobuf==6.31.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pydeck==0.9.1
PyGetWindow==0.0.9
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyMsgBox==1.0.9
pyparsing==3.2.1
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-telegram-bot==22.0
pytweening==1.2.0
pytz==2025.1
pywhatkit==5.4
pywin32-ctypes==0.2.3
referencing==0.36.2
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rfc3986==1.5.0
rpds-py==0.26.0
rsa==4.9
schedule==1.2.2
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
streamlit==1.47.0
tenacity==9.1.2
tkcalendar==1.6.1
toml==0.10.2
tornado==6.5.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.2.0
urllib3==2.3.0
watchdog==6.0.0
Werkzeug==3.0.6
wheel==0.45.1
wikipedia==1.4.0
zipp==3.21.0
